app/src/androidTest/java/com/example/calisthenicsworkout/SkillDatabaseTest.kt

package com.example.calisthenicsworkout.test

import androidx.room.Room
import androidx.test.ext.junit.runners.AndroidJUnit4
import androidx.test.platform.app.InstrumentationRegistry
import com.example.calisthenicsworkout.database.SkillDatabase
import com.example.calisthenicsworkout.database.SkillDatabaseDao
import com.example.calisthenicsworkout.database.entities.Skill
import org.junit.Assert.assertEquals
import org.junit.After
import org.junit.Before
import org.junit.Test
import org.junit.runner.RunWith
import java.io.IOException


@RunWith(AndroidJUnit4::class)
class SleepDatabaseTest {

    private lateinit var sleepDao: SkillDatabaseDao
    private var db = Room.inMemoryDatabaseBuilder(InstrumentationRegistry.getInstrumentation().targetContext, SkillDatabase::class.java).allowMainThreadQueries().build()

    @Before
    fun createDb() {
//        val context = InstrumentationRegistry.getInstrumentation().targetContext
//        // Using an in-memory database because the information stored here disappears when the
//        // process is killed.
//        db = Room.inMemoryDatabaseBuilder(context, SkillDatabase::class.java)
//            // Allowing main thread queries, just for testing.
//            .allowMainThreadQueries()
//            .build()
        sleepDao = db.skillDatabaseDao
    }

    @After
    @Throws(IOException::class)
    fun closeDb() {
        db.close()
    }

    @Test
    @Throws(Exception::class)
    fun insertAndGetSkill() {
        val skill = Skill("Handstand")
        sleepDao.insert(skill)
        val chosenSkill = sleepDao.getSkill(0)
        assertEquals(chosenSkill.id, "Handstand")
    }
}
